IL+RL/requirements.txt
metadrive-simulator
stable-baselines3
imitation
torch
numpy
tensorboard